+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.0.1 (完善PlotDraw)
 
 * 完善PlotDraw
-* 添加部分geometry构造类（Ployline，Arror，Polygon）
+* 添加部分geometry构造类（Ployline，Arror，Polygon, Flag）
 
 ## 0.0.0 (搭建框架)